Refactored projectClass/Location validaiton function into a util function
