+++ infraohjelmointi_api/validators/ProjectValidators/ProjectClassValidator.py
@@ -1,32 +1,12 @@
 from rest_framework.exceptions import ValidationError
+from ..util.is_projectClass_projectLocation_valid import (
+    _is_projectClass_projectLocation_valid,
+)
 
 
 class ProjectClassValidator:
 
     requires_context = True
-
-    def _is_projectClass_projectLocation_valid(
-        self,
-        projectLocation,
-        projectClass,
-    ) -> bool:
-        if projectClass is None or projectLocation is None:
-            return True
-        if projectClass.parent is None or projectClass.parent.parent is None:
-            return True
-        if (
-            "suurpiiri" in projectClass.name.lower()
-            and len(projectClass.name.split()) == 2
-            and (
-                projectLocation.path.startswith(projectClass.name.split()[0])
-                or projectLocation.path.startswith(projectClass.name.split()[0][:-2])
-            )
-        ):
-            return True
-        elif "suurpiiri" not in projectClass.name.lower():
-            return True
-        else:
-            return False
 
     def __call__(self, allFields, serializer) -> None:
         projectClass = allFields.get("projectClass", None)
@@ -42,7 +22,7 @@ class ProjectClassValidator:
         ):
             projectLocation = project.projectLocation
 
-        if not self._is_projectClass_projectLocation_valid(
+        if not _is_projectClass_projectLocation_valid(
             projectLocation=projectLocation, projectClass=projectClass
         ):
             raise ValidationError(
