+++ infraohjelmointi_api/validators/ProjectValidators/ProjectLocationValidator.py
@@ -1,32 +1,12 @@
+from infraohjelmointi_api.validators.util.is_projectClass_projectLocation_valid import (
+    _is_projectClass_projectLocation_valid,
+)
 from rest_framework.exceptions import ValidationError
 
 
 class ProjectLocationValidator:
 
     requires_context = True
-
-    def _is_projectClass_projectLocation_valid(
-        self,
-        projectLocation,
-        projectClass,
-    ) -> bool:
-        if projectClass is None or projectLocation is None:
-            return True
-        if projectClass.parent is None or projectClass.parent.parent is None:
-            return True
-        if (
-            "suurpiiri" in projectClass.name.lower()
-            and len(projectClass.name.split()) == 2
-            and (
-                projectLocation.path.startswith(projectClass.name.split()[0])
-                or projectLocation.path.startswith(projectClass.name.split()[0][:-2])
-            )
-        ):
-            return True
-        elif "suurpiiri" not in projectClass.name.lower():
-            return True
-        else:
-            return False
 
     def __call__(self, allFields, serializer) -> None:
         projectLocation = allFields.get("projectLocation", None)
@@ -41,7 +21,7 @@ class ProjectLocationValidator:
             and project.projectClass is not None
         ):
             projectClass = project.projectClass
-        if not self._is_projectClass_projectLocation_valid(
+        if not _is_projectClass_projectLocation_valid(
             projectLocation=projectLocation, projectClass=projectClass
         ):
             raise ValidationError(
